+++ front/app/modules/commercial/insights/admin/containers/Insights/Edit/InputsTable/index.tsx
@@ -13,6 +13,7 @@ import useInsightsInputs, {
 } from 'modules/commercial/insights/hooks/useInsightsInputs';
 import { IInsightsInputData } from 'modules/commercial/insights/services/insightsInputs';
 import useScanInsightsCategory from 'modules/commercial/insights/hooks/useScanInsightsCategory';
+import useFeatureFlag from 'hooks/useFeatureFlag';
 
 // components
 import { Table, Icon, Box } from 'cl2-component-library';
@@ -139,19 +140,17 @@ const InputsTable = ({
       : // Include only processed input everywhere else
         true;
 
-  const {
-    list: inputs,
-    lastPage,
-    loading,
-    setLoading,
-  } = useInsightsInputs(viewId, {
-    pageNumber,
-    search,
-    sort,
-    processed,
+  const { list: inputs, lastPage, loading, setLoading } = useInsightsInputs(
+    viewId,
+    {
+      pageNumber,
+      search,
+      sort,
+      processed,
 
-    category: selectedCategory,
-  });
+      category: selectedCategory,
+    }
+  );
 
   const { status, progress, triggerScan, onDone } = useScanInsightsCategory(
     viewId,
@@ -159,6 +158,7 @@ const InputsTable = ({
     processed
   );
 
+  const nlpFeatureFlag = useFeatureFlag({ name: 'insights_nlp_flow' });
   // Callbacks and Effects -----------------------------------------------------
 
   // Table Selection
@@ -398,12 +398,8 @@ const InputsTable = ({
       <SearchContainer>
         <SearchInput onChange={onSearch} />
         <Box display="flex" alignItems="center">
-          {inputs.length > 0 && (
-            <Box
-              alignItems="center"
-              mr="16px"
-              display={status === 'isIdle' ? 'flex' : 'none'}
-            >
+          {inputs.length > 0 && nlpFeatureFlag && status === 'isIdle' && (
+            <Box alignItems="center" mr="16px">
               <Button
                 buttonStyle="secondary"
                 textColor={colors.adminTextColor}